SPARK-22181 Addresses Review comments
