+++ sql/core/src/test/resources/sql-tests/results/except.sql.out
@@ -1,12 +1,13 @@
 -- Automatically generated by SQLQueryTestSuite
--- Number of queries: 14
+-- Number of queries: 6
 
 
 -- !query 0
 create temporary view t1 as select * from values
   ("one", 1),
   ("two", 2),
-  ("three", 3)
+  ("three", 3),
+  ("one", NULL)
   as t1(k, v)
 -- !query 0 schema
 struct<>
@@ -18,7 +19,9 @@ struct<>
 create temporary view t2 as select * from values
   ("one", 1),
   ("two", 22),
-  ("one", 5)
+  ("one", 5),
+  ("one", NULL),
+  (NULL, 5)
   as t2(k, v)
 -- !query 1 schema
 struct<>
@@ -27,103 +30,36 @@ struct<>
 
 
 -- !query 2
-create temporary view t3 as select * from t1 where v <> 1 and v <> 2
+SELECT * FROM t1 EXCEPT SELECT * FROM t2
 -- !query 2 schema
-struct<>
+struct<k:string,v:int>
 -- !query 2 output
-
+three	3
+two	2
 
 
 -- !query 3
-create temporary view t4 as select * from t1 where v <> 2 and v >= 3
+SELECT * FROM t1 EXCEPT SELECT * FROM t1 where v <> 1 and v <> 2
 -- !query 3 schema
-struct<>
+struct<k:string,v:int>
 -- !query 3 output
-
+one	1
+one	NULL
+two	2
 
 
 -- !query 4
-create temporary view t5 as select * from t2 where v >= 1 and v <> 22
+SELECT * FROM t1 where v <> 1 and v <> 22 EXCEPT SELECT * FROM t1 where v <> 2 and v >= 3
 -- !query 4 schema
-struct<>
+struct<k:string,v:int>
 -- !query 4 output
-
+two	2
 
 
 -- !query 5
-SELECT * FROM t1
+SELECT * FROM t2 where v >= 1 and v <> 22 EXCEPT SELECT * FROM t1
 -- !query 5 schema
 struct<k:string,v:int>
 -- !query 5 output
-one	1
-three	3
-two	2
-
-
--- !query 6
-SELECT * FROM t2
--- !query 6 schema
-struct<k:string,v:int>
--- !query 6 output
-one	1
-one	5
-two	22
-
-
--- !query 7
-SELECT * FROM t3
--- !query 7 schema
-struct<k:string,v:int>
--- !query 7 output
-three	3
-
-
--- !query 8
-SELECT * FROM t4
--- !query 8 schema
-struct<k:string,v:int>
--- !query 8 output
-three	3
-
-
--- !query 9
-SELECT * FROM t5
--- !query 9 schema
-struct<k:string,v:int>
--- !query 9 output
-one	1
-one	5
-
-
--- !query 10
-SELECT * FROM t1 EXCEPT SELECT * FROM t2
--- !query 10 schema
-struct<k:string,v:int>
--- !query 10 output
-three	3
-two	2
-
-
--- !query 11
-SELECT * FROM t1 EXCEPT SELECT * FROM t3
--- !query 11 schema
-struct<k:string,v:int>
--- !query 11 output
-one	1
-two	2
-
-
--- !query 12
-SELECT * FROM t3 EXCEPT SELECT * FROM t4
--- !query 12 schema
-struct<k:string,v:int>
--- !query 12 output
-
-
-
--- !query 13
-SELECT * FROM t5 EXCEPT SELECT * FROM t1
--- !query 13 schema
-struct<k:string,v:int>
--- !query 13 output
+NULL	5
 one	5
